+++ frontend/src/components/Footer.tsx
@@ -1,15 +1,13 @@
-
 const Footer = () => {
-    return (
-        <footer className="mt-12 mb-5 flex justify-between text-gray-500">
-            <div className="flex gap-6">
-            <p className="cursor-pointer hover:text-white transition">Privacy</p>
-            <p className="cursor-pointer hover:text-white transition">Terms</p>
-            </div>
-            <p>© 2024 Duel-Learn Inc.</p>
-        </footer>
-    );
-
+  return (
+    <footer className="mt-12 mb-5 flex justify-between text-gray-500 px-8">
+      <div className="flex gap-6">
+        <p className="cursor-pointer hover:text-white transition">Privacy</p>
+        <p className="cursor-pointer hover:text-white transition">Terms</p>
+      </div>
+      <p>© 2024 Duel-Learn Inc.</p>
+    </footer>
+  );
 };
 
 export default Footer;
