--- frontend/src/pages/dashboard/shop/BuyPremimun.css
@@ -1,8 +1,0 @@
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
-}
