+++ frontend/src/pages/dashboard/shop/ShopPage.tsx
@@ -65,7 +65,7 @@ const Shop = () => {
             </p>
             <button
               className="mt-4 px-10 py-2 text-[19px] bg-white text-[#9F87E5] rounded-full font-bold"
-              onClick={() => navigate("/dashboard/shop/buy-premium-account")}
+              onClick={() => navigate("/dashboard/buy-premium-account")}
             >
               TRY IT NOW
             </button>
